+++ internal/modules/iot_device/iot_device.model.go
@@ -14,7 +14,7 @@ type IoTDevice struct {
 	MacAddress string       `gorm:"primaryKey;type:varchar(50)"`
 	DeviceName *string      `gorm:"type:varchar(50)"`
 	Status     DeviceStatus `gorm:"type:enum('ACTIVE','INACTIVE','ERROR');default:'ACTIVE';not null"`
-	LotID      *uint        `gorm:"index"`
+	LotID      *uint64      `gorm:"index"`
 	LastSeen   *time.Time
 	CreatedAt  time.Time `gorm:"autoCreateTime"`
 	UpdatedAt  time.Time `gorm:"autoUpdateTime"`
