+++ internal/modules/parking_session/parking_session.handler.go
@@ -86,7 +86,7 @@ func (h *Handler) GetByDate(c *gin.Context) {
 		return
 	}
 
-	userID, ok := userIDValue.(uint)
+	userID, ok := userIDValue.(uint64)
 	if !ok {
 		c.Error(appErrors.NewUnauthorized("Unauthorized"))
 		return
